Add Dockerfiles with development versions of Gerrit
This project supported only docker images containing release builds of
Gerrit. To test changes on Gerrit, it would be useful to be able to
build docker images containing Gerrit with this change.

This change adds Dockerfiles based on Ubuntu/Centos that allow to build
docker images containing Gerrit at the state of a given git reference.
This is done by using the Bazel build slave docker image maintained in
the gerrit-ci-scripts project [1] to build Gerrit. The git reference can
be set at build time by giving the `GERRIT_REF` build argument to the
`docker build`-command. The build artifact is then copied into the
image, which is used to run Gerrit, as part of a multistage build.

[1] https://gerrit-review.googlesource.com/admin/repos/gerrit-ci-scripts

Change-Id: I3d77ac0e2fad44e24e48e10ac910df32bcd0cdbd