Add Dockerfiles with development versions of Gerrit
This project supported only docker images containing release builds of
Gerrit. To test changes on Gerrit, it would be useful to be able to
build docker images containing Gerrit with this change.

This change adds Dockerfiles based on Ubuntu/Centos that allow to build
docker images containing Gerrit run from a custom *.war-file. The
*.war-file is provided during build time by setting the
`GERRIT_WAR_URL`-build argument to a URL pointing to the respective
file.

Change-Id: I3d77ac0e2fad44e24e48e10ac910df32bcd0cdbd