README.md: fix the database type from ReviewDB to noteDb
Actually, this readme is used for newcommers.
According the gerrit release note and docs from gerritcodereview.com,
noteDb is the only one database used instead of ReviewDB.
So, change this readme for newcommers. Or, database migration will get
errors when up gerrit docker container and gerrit site cannot be accessed
with a error code "500".

Signed-off-by: Daniel Tse <learningendless@gmail.com>
Change-Id: Ib72187b05f49937fbe164a127b15c77400a27bac