Run init during docker build
The Gerrit init step is responsible for extracting the gerrit startup
script, the service wrapper and the installation of plugins.
Running it only at runtime would not allow existing sites to work
properly or even to be migrated.

Introduce a 'init' parameter to the entrypoint.sh so that it can be
invoked at build time and from the docker-compose.yaml for upgrading
existing sites.

Move also the java.security.egd=file:/dev/./urandom setting to
entrypoint.sh, so to make sure that gets executed also on externally
mounted configuration files.

Change-Id: I903b2ed49278f5c929d63f3e65ac1bb96b1de081