Allow to use init as a param in docker run
The problem is that the parameter for docker run added at the end is not
passed down to the entry point script as it is right now. This is easly
testable - just run the gerritreview image with init over and over and
it will always start gerrit. Applying the patch the same command will
call init.

Change-Id: I19ed2b33b388026f9a4fbbc6ebde6bb4cfc42111