Avoid warning on illegal reflective access operation
Java 9 introduced the modules system which enforces the strong
encapsulation of packages inside a module.

The Guice (assisted inject) and Guava (IoUtils) libraries have
been developed on Java 8 where the module system did not exist
and therefore are making access to internal methods of java.net
and java.lang.invoke packages that are not allowed anymore.

Silence for now the warnings, until Guice and Guava will be
upgraded to a newer version more compliant with the Java 9
modules strong encapsulation.

Bug: Issue 12639
Change-Id: I47d0ca65b2f62cdb1397f47f84e0f51a41c80c08