Add helper script to perform multi-platform images builds
The `build_multiplatform.sh` file was introduced to automate the
multi-platform images build process.
One calls

  ./build_multiplatform.sh

to build both ubuntu and almalinux images for both `amd64` and `arm64`
platforms and load them to the local docker registry.

One calls:

  ./build_multiplatform.sh --push

in order to push them to the dockerhub's gerritcodereview account.

Notes:
* script takes care of multi-platform builder creation and clean-up
* almalinux is additionlly tagged as the default release image
* Gerrit installation packages need to be build and published using the
  changes made in Iedcaf6511 as otherwise image build will fail for
  arm64 platform

Change-Id: I5baaed04ec8a82652c46d1f4ef151928690d7cfb
Depends-On: Iedcaf6511fb091775c1f2d0fb1596531efbba7e1