Explain how adding an external-mount may change first-run behaviour
It is common for /var/gerrit/etc to be externally mounted to an
initially empty directory on the host.  If that is done, then the
special copy of gerrit.config that comes from the deb/rpm package
won't be found when gerrit starts.  Gerrit will generate a new config,
but it will not contain the special settings present in the one from
the deb/rpm package and the start-up behaviour will be different
(e.g. the "admin" user will not be automatically created, and
plugin-manager will not show the welcome screen).  Add a note
explaining this.  Also, add a link to help readers understand that
"out-of-the-box" refers to a plugin.

Change-Id: I2aac83fe822e629a3409136f52d87232dfce5661