Migrate to eclipse-temurin based JDK docker image
As a result ubuntu image creation is faster (as JDK package download and
install is no longer executed) and the image is substantially smaller.
On M1 it is correspondingly:
* ubuntu based 867MB
* temurin based 643MB

Note that temurin does everything that is needed in order to install JDK
but doesn't update alternatives for java which is detected during the
gerrit install step hence it is the only extra step that had to be
added.

Bug: Issue 303743162
Change-Id: I015311ca4da85b6ce6bb090a5b9baebcae715af0